some modularization, towards the goal of being able to generate mesh from reading the .csv file
