+++ HH_shape_opt/generate_mesh/generate_initial_mesh.py
@@ -1,0 +1,133 @@
+import meshio
+import matplotlib.pyplot as plt
+import numpy as np
+import os
+import gmsh
+
+from .mesh_util import SIDE_WALL_MARKER, RECEIVER_EDGE_MARKER, OBSTACLE_MARKER, OBSTACLE_OPT_MARKER, DOMAIN_MARKER
+
+
+def generate_square_with_rect_obstacle_mesh(
+    width=1.0, height=1.0, rect_w=0.4, rect_h=0.2, mesh_size=0.05,
+    output_name="square_with_rect_obstacle",
+    n_points_bottom=100, n_points_rect_bottom=40,
+    use_opt_marker=False
+):
+
+    gmsh.initialize()
+    gmsh.clear()
+    gmsh.model.add("square_with_rect_obstacle")
+
+    # Outer square points
+    p1 = gmsh.model.geo.addPoint(0, 0, 0, mesh_size)         # Bottom-left
+    p2 = gmsh.model.geo.addPoint(width, 0, 0, mesh_size)     # Bottom-right
+    p3 = gmsh.model.geo.addPoint(width, height, 0, mesh_size)  # Top-right
+    p4 = gmsh.model.geo.addPoint(0, height, 0, mesh_size)    # Top-left
+
+    # Outer square lines
+    l1 = gmsh.model.geo.addLine(p1, p2)  # Bottom
+    l2 = gmsh.model.geo.addLine(p2, p3)  # Right
+    l3 = gmsh.model.geo.addLine(p3, p4)  # Top
+    l4 = gmsh.model.geo.addLine(p4, p1)  # Left
+
+    # Use TransfiniteCurve for bottom wall discretization
+    gmsh.model.geo.mesh.setTransfiniteCurve(l1, n_points_bottom)
+
+    # Rectangle obstacle center
+    cx, cy = width/2, height/2
+    rx1 = cx - rect_w/2
+    rx2 = cx + rect_w/2
+    ry1 = cy - rect_h/2
+    ry2 = cy + rect_h/2
+
+    # Rectangle obstacle points (counterclockwise)
+    rp1 = gmsh.model.geo.addPoint(rx1, ry1, 0, mesh_size)
+    rp2 = gmsh.model.geo.addPoint(rx2, ry1, 0, mesh_size)
+    rp3 = gmsh.model.geo.addPoint(rx2, ry2, 0, mesh_size)
+    rp4 = gmsh.model.geo.addPoint(rx1, ry2, 0, mesh_size)
+
+    # Rectangle obstacle lines
+    rl1 = gmsh.model.geo.addLine(rp1, rp2)
+    rl2 = gmsh.model.geo.addLine(rp2, rp3)
+    rl3 = gmsh.model.geo.addLine(rp3, rp4)
+    rl4 = gmsh.model.geo.addLine(rp4, rp1)
+    rect_lines = [rl1, rl2, rl3, rl4]
+
+    gmsh.model.geo.mesh.setTransfiniteCurve(rl1, n_points_rect_bottom)
+
+    # Curve loops
+    outer_loop = gmsh.model.geo.addCurveLoop([l1, l2, l3, l4])
+    rect_loop = gmsh.model.geo.addCurveLoop(rect_lines)
+
+    # Plane surface with rectangle obstacle
+    surface = gmsh.model.geo.addPlaneSurface([outer_loop, rect_loop])
+
+    gmsh.model.geo.synchronize()
+
+    # Physical groups
+    gmsh.model.addPhysicalGroup(1, [l1], RECEIVER_EDGE_MARKER, "bottom_wall")
+    gmsh.model.addPhysicalGroup(
+        1, [l2, l3, l4], SIDE_WALL_MARKER, "outer_walls")
+    if use_opt_marker:
+        # Mark the bottom of the obstacle separately for optimization
+        gmsh.model.addPhysicalGroup(
+            1, [rl1], OBSTACLE_OPT_MARKER, "obstacle_opt_boundary")
+        # Mark the rest of the obstacle
+        gmsh.model.addPhysicalGroup(
+            1, [rl2, rl3, rl4], OBSTACLE_MARKER, "rect_obstacle_boundary")
+    else:
+        # Mark the entire obstacle with one marker
+        gmsh.model.addPhysicalGroup(
+            1, rect_lines, OBSTACLE_MARKER, "rect_obstacle_boundary")
+    gmsh.model.addPhysicalGroup(2, [surface], DOMAIN_MARKER, "domain")
+
+    # Generate mesh
+    gmsh.model.mesh.generate(2)
+    gmsh.option.setNumber("Mesh.MshFileVersion", 2.2)
+    gmsh.write(f"{output_name}.msh")
+    gmsh.finalize()
+    return f"{output_name}.msh"
+
+
+
+
+if __name__ == "__main__":
+    print("Generating square with hole mesh...")
+
+    c = 299792458
+    freq_max = 5e9  # 5GHz
+
+    # Parameters
+    wavelength = c / freq_max  # Physical wavelength
+    mesh_size = wavelength / 5
+
+    """
+    mesh_file = generate_square_with_cos_perturbed_rect_obstacle_mesh(
+        width=1.0, height=1.0, rect_w=0.4, rect_h=0.2, mesh_size=mesh_size,
+        output_name="meshes/square_with_perturbed_rect_obstacle",
+        n_points_bottom=100, n_points_rect_bottom=100,
+        perturb_amplitude=0.01, perturb_frequency=1.5
+    )
+    """
+
+    """
+    mesh_file = generate_square_with_sin_perturbed_rect_obstacle_mesh(
+        width=1.0, height=1.0, rect_w=0.4, rect_h=0.2, mesh_size=mesh_size,
+        output_name="square_with_halfsin2_perturbed_rect_obstacle",
+        n_points_bottom=100, n_points_rect_bottom=100,
+        perturb_amplitude=0.02, perturb_frequency=0.5
+    )
+    """
+
+    """
+    mesh_file =  generate_square_with_rect_obstacle_mesh(
+    width=1.0, height=1.0, rect_w=0.4, rect_h=0.2, mesh_size=mesh_size,
+    output_name="meshes/square_with_rect_obstacle_opt",
+    n_points_bottom=100, n_points_rect_bottom=100,
+    use_opt_marker = True
+    )
+    """
+
+    """
+    convert_msh_to_xdmf(mesh_file)
+    """
