+++ opt_DG0/opt_script.py
@@ -10,11 +10,7 @@ import os
 import gmsh
 import matplotlib.pyplot as plt
 
-from HH_shape_opt.mesh_generation import obstacle_marker, side_wall_marker, bottom_wall_marker, obstacle_opt_marker
-from HH_shape_opt.initialize_opt import MeshUtil
-from HH_shape_opt.helmholtz_solve import forward_solve, load_forward_simulation_data_bottomwall, IncidentWaveSetup, plane_wave
-from HH_shape_opt.process_result import save_optimization_result, calculate_magnitude_and_phase_error
-from HH_shape_opt.visualize import plot_mesh_deformation_from_result, plot_projected_errors
+from HH_shape_opt import *
 
 # Ensure this can be run from root dir
 script_dir = os.path.dirname(os.path.abspath(__file__))
@@ -28,9 +24,9 @@ inc_wave_setup = IncidentWaveSetup(frequency, plane_wave)
 measurement_data_file_path = "measurements/matlab_measurements_sin0.5_top_perturbed.csv"
 msh_file_path = "meshes/square_with_rect_obstacle.msh"
 markers_dict = {
-    "obstacle": obstacle_marker,
-    "side_wall": side_wall_marker,
-    "bottom_wall": bottom_wall_marker,
+    "obstacle": OBSTACLE_MARKER, # Markers importee from our mesh generation module
+    "side_wall": SIDE_WALL_MARKER,
+    "bottom_wall": RECEIVER_EDGE_MARKER,
     "obstacle_opt": None
 }
 obstacle_stiffness = 25
@@ -54,7 +50,7 @@ h_V.rename("Volume extension of h", "")
 ##########################
 
 # Solve the forward problem
-u_tot_mag_dg0, u_tot_re_projected, u_tot_im_projected, ds_bottom, V_DG0 = forward_solve(
+u_tot_mag_dg0, u_tot_re_projected, u_tot_im_projected, ds_receiver, V_DG0 = forward_solve(
     h, inc_wave_setup, initial_guess_mesh_util)
 
 # Load the reference data in the same function space as the projected result of the forward solve
@@ -62,7 +58,7 @@ u_ref_dg0, _ = load_forward_simulation_data_bottomwall(
     measurement_data_file_path, V_DG0)
 
 J = assemble(
-    (inner(u_tot_mag_dg0 - u_ref_dg0, u_tot_mag_dg0 - u_ref_dg0) * ds_bottom))
+    (inner(u_tot_mag_dg0 - u_ref_dg0, u_tot_mag_dg0 - u_ref_dg0) * ds_receiver))
 Jhat = ReducedFunctional(J, Control(h))
 
 dJdh = Jhat.derivative()
